+++ example/lib/main.dart
@@ -44,7 +44,7 @@ class _MyAppState extends State<MyApp> {
     final config = ldk.Config(
         storageDirPath: nodePath,
         esploraServerUrl: esploraUrl,
-        network: ldk.Network.Regtest,
+        network: ldk.Network.regtest,
         listeningAddress: address,
         defaultCltvExpiryDelta: 144);
     return config;
@@ -58,12 +58,11 @@ class _MyAppState extends State<MyApp> {
   initAliceNode() async {
     final aliceConfig = await initLdkConfig(
         'alice', const ldk.SocketAddr(ip: "0.0.0.0", port: 3006));
-    ldk.Builder aliceBuilder = ldk.Builder.fromConfig(config: aliceConfig)
-        .setEntropyBip39Mnemonic(
-            mnemonic:
-                'puppy interest whip tonight dad never sudden response push zone pig patch');
+    ldk.Builder aliceBuilder = ldk.Builder.fromConfig(config: aliceConfig);
+    aliceBuilder.setEntropyBip39Mnemonic(
+        mnemonic:
+            'certain sense kiss guide crumble hint transfer crime much stereo warm coral');
     aliceNode = await aliceBuilder.build();
-
     await aliceNode.start();
     final res = await aliceNode.nodeId();
     setState(() {
@@ -76,7 +75,6 @@ class _MyAppState extends State<MyApp> {
     final bobConfig = await initLdkConfig(
         "bob", const ldk.SocketAddr(ip: "0.0.0.0", port: 8077));
     ldk.Builder bobBuilder = ldk.Builder.fromConfig(config: bobConfig);
-
     bobNode = await bobBuilder.build();
     await bobNode.start();
     final res = await bobNode.nodeId();
@@ -123,7 +121,7 @@ class _MyAppState extends State<MyApp> {
 
   Future<ldk.PaymentDetails?> listPayments(bool printPayments) async {
     final res = await aliceNode.listPaymentsWithFilter(
-        paymentDirection: ldk.PaymentDirection.Outbound);
+        paymentDirection: ldk.PaymentDirection.outbound);
     if (res.isNotEmpty) {
       if (printPayments) {
         if (kDebugMode) {