+++ lib/ldk_node.dart
@@ -1,12 +1,13 @@
 export './src/generated/bridge_definitions.dart'
     hide
-        Rust,
+        RustLdkNode,
         Event_ChannelClosed,
         Event_ChannelReady,
         Event_PaymentFailed,
         Event_PaymentReceived,
         Event_PaymentSuccessful,
         Event_ChannelPending,
+        NodePointer,
         WalletEntropySource_SeedBytes,
         WalletEntropySource_Bip39Mnemonic,
         WalletEntropySource_SeedFile;